feat: refactor permissions and extras modules for improved entity relationships and database integration

+++ src/entities/extras.entity.ts
@@ -12,7 +12,7 @@ export class ExtrasEntity {
   @Column("double", { name: "item_price", precision: 8, scale: 2 })
   itemPrice: number;
 
-  @Column("decimal", {
+  @Column("double", {
     name: "commission",
     precision: 8,
     scale: 2,
